+++ src/main/java/net/blueshell/api/business/news/NewsDTO.java
@@ -119,8 +119,8 @@ public class NewsDTO {
 
     //TODO: fix this when we work with POST (the id part)
     public News toNews() {
-        User author = userDao.getById(Integer.parseInt(creatorId));
-        User last_editor = userDao.getById(Integer.parseInt(lastEditorId));
+        User author = userDao.getById(Long.parseLong(creatorId));
+        User last_editor = userDao.getById(Long.parseLong(lastEditorId));
         Timestamp posted_at = Timestamp.valueOf(this.postedAt);
         return new News(author, newsType, title, content, posted_at);
     }
